refactoring to remove vaccine only codes
