+++ __tests__/Verify-LabResult-Test.ts
@@ -4,7 +4,7 @@ import labResultValide from '../src/services/fhir/recordValidator/labResultValid
 import R4Observation from '../src/services/fhir/recordValidator/labResultValidator/systems/R4Observation'
 import fhirData from './fixtures/fhirLabResult.json'
 import _ from 'lodash'
-import { getAcceptedSystemCodes } from '../src/services/helpers/getVaccineCodesHash'
+import { getAcceptedSystemCodes } from '../src/services/helpers/getFHIRCodes'
 it('Verifies incorrectly', async () => {
   var res = labResultValide([])
   expect(res).toEqual(false)
@@ -15,10 +15,9 @@ it('Verifies incorrectly', async () => {
   expect(res).toEqual(false)
 })
 
-it('Verifies Observation', ()=> {
+it('GettingCodeHash', ()=> {
   const observation = new R4Observation()
   var res = observation.isSystem( fhirData[1] )
   expect(res).toEqual(true)
-
-  
 })
+
