+++ src/services/fhir/recordParser/ImmunizationRecordParser.ts
@@ -1,5 +1,5 @@
 import { formatVaccinationDate } from '../../../utils/utils'
-import { getVaccineCodesHash, getAcceptedVaccineCodes } from '../../helpers/getVaccineCodesHash'
+import { getVaccineCodesHash, getAcceptedVaccineCodes } from '../../helpers/getFHIRCodes'
 import { ResourceType } from '../fhirTypes'
 import { isResourceType } from '../fhirBundle'
